Move schema and product error modultes into store module
This is so that these store modules are consistent with the structure of
the other store modules.
