+++ sdk/src/grid_db/products/store/diesel/operations/update_product.rs
@@ -15,7 +15,8 @@
 use super::ProductStoreOperations;
 
 use crate::grid_db::products::{
-    error::ProductStoreError, store::diesel::schema::product_property_value, MAX_COMMIT_NUM,
+    store::{diesel::schema::product_property_value, error::ProductStoreError},
+    MAX_COMMIT_NUM,
 };
 use diesel::{dsl::update, prelude::*};
