+++ sdk/src/grid_db/products/store/mod.rs
@@ -14,8 +14,9 @@
 
 #[cfg(feature = "diesel")]
 pub mod diesel;
+pub mod error;
 
-use crate::grid_db::products::error::ProductStoreError;
+pub use error::ProductStoreError;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Product {
