+++ sdk/src/grid_db/schemas/mod.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod error;
 pub mod store;
 
 pub const MAX_COMMIT_NUM: i64 = i64::MAX;
