+++ sdk/src/grid_db/schemas/store/error.rs
